Deployment

+++ WebEng.Spares/Models/OEM.cs
@@ -13,7 +13,7 @@ namespace WebEng.ReplacementParts.Models
         public string OEMNumber { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public long Weight { get; set; }
+        public double Weight { get; set; }
         public string PictureUrl { get; set; }
     }
 }
