+++ WebEng.Spares/Models/SparePart.cs
@@ -17,8 +17,8 @@ namespace WebEng.ReplacementParts.Models
         public long ManufacturerKey { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public long Price { get; set; }
-        public long Weight { get; set; }
+        public double Price { get; set; }
+        public double Weight { get; set; }
         public long Available { get; set; }
         public string PictureUrl { get; set; }
     }
